refactor: Rename contants file to `consts`
